add tests for the style macro merging
